I fix the accelerate so that cars can move in constant speed after a certain amount of time
I create a move method. Christine, please put your turning method in accordingly

+++ src/traffic_congestion_simulator/TCSConstant.java
@@ -86,5 +86,5 @@ public interface TCSConstant {
     
     public final int NUMOFINTERSECTION = 0;
     
-    public final int AUTOMATEDFINALVELOCITY = 10;
+    public final double AUTOMATEDFINALVELOCITY = 10.0;
 }
